+++ src/dcos_e2e/backends/_docker/__init__.py
@@ -3,7 +3,6 @@ Helpers for creating and interacting with clusters on Docker.
 """
 
 import inspect
-import logging
 import os
 import socket
 import subprocess
@@ -20,7 +19,7 @@ from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
 
-from dcos_e2e._common import run_subprocess
+from dcos_e2e._common import get_logger, run_subprocess
 from dcos_e2e.backends._base_classes import ClusterBackend, ClusterManager
 from dcos_e2e.distributions import Distribution
 from dcos_e2e.docker_storage_drivers import DockerStorageDriver
@@ -30,8 +29,7 @@ from dcos_e2e.node import Node
 from ._containers import start_dcos_container
 from ._docker_build import build_docker_image
 
-logging.basicConfig(level=logging.DEBUG)
-LOGGER = logging.getLogger(__name__)
+LOGGER = get_logger(__name__)
 
 
 def _write_key_pair(public_key_path: Path, private_key_path: Path) -> None: